src/requirements/base.txt
# Core
Flask==1.0.2
psycopg2>=2.6,<2.7
Flask-SQLAlchemy>=2.3.1,<2.3.2
Flask-Migrate>=2.1.1,<2.1.2
Flask-Script==2.0.6
flask-restful==0.3.6
flask-bcrypt==0.7.1
flask-rest-jsonapi==0.19.1
email_validator==1.0.3
flask-jwt==0.3.2
mysqlclient==1.3.13

# Server
gunicorn==19.6.0
